requirements.txt
streamlit
requests
asyncio
pillow
